+++ tree/memslot.cpp
@@ -29,7 +29,7 @@ void MemSlotVector::init(PhyloTree *tree, int num_slot) {
     if (Params::getInstance().lh_mem_save != LM_MEM_SAVE)
         return;
     reserve(num_slot+2);
-    resize(num_slot);
+    resize(static_cast<size_t>(num_slot));
     size_t lh_size = tree->getPartialLhSize();
     size_t scale_size = tree->getScaleNumSize();
     reset();
@@ -76,7 +76,7 @@ void MemSlotVector::addSpecialNei(PhyloNeighbor *nei) {
     ms.partial_lh = nei->partial_lh;
     ms.scale_num = nei->scale_num;
     push_back(ms);
-    nei_id_map[nei] = size()-1;
+    nei_id_map[nei] = static_cast<int>(size())-1;
 }
 
 void MemSlotVector::eraseSpecialNei() {
